:recycle: refactor: AdReward 패키지명 변경
