--- src/main/java/tipitapi/drawmytoday/ad/domain/AdType.java
@@ -1,5 +1,0 @@
-package tipitapi.drawmytoday.ad.domain;
-
-public enum AdType {
-    VIDEO
-}
