+++ src/main/java/tipitapi/drawmytoday/adreward/domain/AdReward.java
@@ -1,4 +1,4 @@
-package tipitapi.drawmytoday.ad.domain;
+package tipitapi.drawmytoday.adreward.domain;
 
 import java.time.LocalDateTime;
 import javax.persistence.Entity;
