+++ src/main/java/tipitapi/drawmytoday/adreward/domain/AdType.java
@@ -1,0 +1,5 @@
+package tipitapi.drawmytoday.adreward.domain;
+
+public enum AdType {
+    VIDEO
+}
