+++ src/main/java/tipitapi/drawmytoday/adreward/repository/AdRewardRepository.java
@@ -1,11 +1,11 @@
-package tipitapi.drawmytoday.ad.repository;
+package tipitapi.drawmytoday.adreward.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import tipitapi.drawmytoday.ad.domain.AdReward;
+import tipitapi.drawmytoday.adreward.domain.AdReward;
 
 public interface AdRewardRepository extends JpaRepository<AdReward, Long> {
