+++ src/main/java/tipitapi/drawmytoday/adreward/service/AdRewardService.java
@@ -1,11 +1,11 @@
-package tipitapi.drawmytoday.ad.service;
+package tipitapi.drawmytoday.adreward.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import tipitapi.drawmytoday.ad.domain.AdReward;
-import tipitapi.drawmytoday.ad.domain.AdType;
-import tipitapi.drawmytoday.ad.repository.AdRewardRepository;
+import tipitapi.drawmytoday.adreward.domain.AdReward;
+import tipitapi.drawmytoday.adreward.domain.AdType;
+import tipitapi.drawmytoday.adreward.repository.AdRewardRepository;
 import tipitapi.drawmytoday.user.domain.User;
 import tipitapi.drawmytoday.user.service.ValidateUserService;
